+++ gear/form_qc_checker/src/python/form_qc_app/processor.py
@@ -22,7 +22,7 @@ from outputs.error_logger import (
     MetadataCleanupFlag,
     update_error_log_and_qc_metadata,
 )
-from outputs.error_models import JSONLocation
+from outputs.error_models import JSONLocation, VisitKeys
 from outputs.error_writer import ListErrorWriter
 from outputs.errors import (
     empty_field_error,
@@ -259,7 +259,12 @@ class JSONFileProcessor(FileProcessor):
             # has a failed previous visit
             if failed_visit.visitdate < visitdate:
                 self._error_writer.write(
-                    previous_visit_failed_error(failed_visit.filename)
+                    previous_visit_failed_error(
+                        prev_visit=failed_visit.filename,
+                        visit_keys=VisitKeys.create_from(
+                            record=self.__input_record, date_field=self._date_field
+                        ),
+                    )
                 )
                 return "DIFFERENT"
 
@@ -304,7 +309,14 @@ class JSONFileProcessor(FileProcessor):
                 found_all = False
 
         if not found_all:
-            self._error_writer.write(empty_field_error(empty_fields))
+            self._error_writer.write(
+                empty_field_error(
+                    field=empty_fields,
+                    visit_keys=VisitKeys.create_from(
+                        record=input_data, date_field=self._date_field
+                    ),
+                )
+            )
             return None
 
         subject_lbl = input_data[self._pk_field]
@@ -316,7 +328,13 @@ class JSONFileProcessor(FileProcessor):
             )
             log.error(message)
             self._error_writer.write(
-                system_error(message, JSONLocation(key_path=self._pk_field))
+                system_error(
+                    message=message,
+                    error_location=JSONLocation(key_path=self._pk_field),
+                    visit_keys=VisitKeys.create_from(
+                        record=input_data, date_field=self._date_field
+                    ),
+                )
             )
             return None