+++ gear/identifier_lookup/src/python/identifier_app/main.py
@@ -15,7 +15,7 @@ from identifiers.model import IdentifierObject, clean_ptid
 from inputs.csv_reader import CSVVisitor, read_csv
 from keys.keys import FieldNames
 from outputs.error_logger import update_error_log_and_qc_metadata
-from outputs.error_models import FileError
+from outputs.error_models import FileError, VisitKeys
 from outputs.error_writer import ListErrorWriter
 from outputs.errors import (
     identifier_error,
@@ -69,7 +69,11 @@ class NACCIDLookupVisitor(CSVVisitor):
         self.__gear_name = gear_name
         self.__header: Optional[List[str]] = None
         self.__writer: Optional[CSVWriter] = None
-        self.__validator = CenterValidator(center_id=adcid, error_writer=error_writer)
+        self.__validator = CenterValidator(
+            center_id=adcid,
+            date_field=module_configs.date_field,
+            error_writer=error_writer,
+        )
         self.__misc_errors = misc_errors
 
     def __get_writer(self) -> CSVWriter:
@@ -124,6 +128,8 @@ class NACCIDLookupVisitor(CSVVisitor):
         Returns:
           True if there is a NACCID for the PTID, False otherwise
         """
+
+        # processing a new row, clear previous errors if any
         self.__error_writer.clear()
 
         # check for valid ADCID and PTID
@@ -195,8 +201,8 @@ class NACCIDLookupVisitor(CSVVisitor):
             self.__misc_errors.append(
                 unexpected_value_error(
                     field=f"{FieldNames.PTID} or {self.__module_configs.date_field}",
-                    value=" ",
-                    expected=" ",
+                    value="",
+                    expected="",
                     message=message,
                 )
             )
@@ -216,7 +222,14 @@ class NACCIDLookupVisitor(CSVVisitor):
                 "Failed to update error log for visit "
                 f"{input_record[FieldNames.PTID]}_{input_record[self.__module_configs.date_field]}"
             )
-            self.__misc_errors.append(system_error(message))
+            self.__misc_errors.append(
+                system_error(
+                    message=message,
+                    visit_keys=VisitKeys.create_from(
+                        record=input_record, date_field=self.__module_configs.date_field
+                    ),
+                )
+            )
             return False
 
         return True
@@ -302,7 +315,9 @@ class CenterLookupVisitor(CSVVisitor):
             raise GearExecutionError(f"Lookup of {naccid} failed: {error}") from error
 
         if not identifier:
-            self.__error_writer.write(identifier_error(line=line_num, value=naccid))
+            self.__error_writer.write(
+                identifier_error(line=line_num, value=naccid, field=FieldNames.NACCID)
+            )
             return False
 
         row[FieldNames.ADCID] = identifier.adcid